fix(execexpr): ARRAY[] of >=65536 elements silently returned an empty array
ArrayExprStep.nelems was u16 and compile.rs stored `nelems as u16`, so an
ARRAY[] literal with 65536 elements wrapped to 0 and evaluated to an EMPTY
array with NO error raised:

    select array_length(ARRAY[ ...65536 elements... ], 1)
      C:      65536
      pgrust: NULL

65535 was correct, which is what made it hide. C's execExpr.h keeps this count
in an `int` — there is no 16-bit ceiling to honour — so the fix is to widen,
not to add a limit. Found by pgjdbc's PreparedStatement64KBindsTest and
reproduced in bare psql, so it is an engine bug, not a driver artifact.

Silently wrong results are the worst failure class: a crash is loud, this
just hands back the wrong answer. Two siblings carried the same u16 for the
same kind of unbounded count and are widened with it:
  * RowExprStep.nelems  (C rowexpr.nelems is `int`)
  * MinMax.nelems       (GREATEST/LEAST over an int-length arg List)
Audited the rest: the remaining `.len() as u16` sites are WAL-record and
page-item counts (nbtree, hash, gist, heap), which are genuinely uint16 in C's
WAL structs and bounded by page size — those are correct.

Regression test walks the boundary at 65535/65536/65537, asserting both the
count the compiled step carries and the length of the array actually built.
Proven non-vacuous: reintroducing `as u16` makes it fail with "wrapped at
65536 elements".

Also in this commit, the elog source-line fixes from the same driver sweep:
NoticeResponse/ErrorResponse now always emit PG_DIAG_SOURCE_LINE when a
location is present, because C's ereport macros always pass __LINE__ and
clients rely on it — asyncpg does an unguarded del dct['server_source_line'],
which raised KeyError inside its notice callback and hung the connection. A 0
there means "this ported site has no C line yet" (greppable) rather than a
fabricated number; pl/pgsql RAISE now carries its real one, pl_exec.c:3909,
verified against the REL_18_3 source.

+++ crates/backend/executor/execexpr/src/steps.rs
@@ -145,7 +145,11 @@ pub enum Step {
     // `frame` is an argless FuncFrame carried only for its armed result mcx.
     ArrayExprStep {
         elems: NonNull<NullableDatum>,
-        nelems: u16,
+        // C execExpr.h arrayexpr.nelems is `int`: ARRAY[] element count is
+        // NOT bounded by any 16-bit limit. This was u16, so a literal with
+        // 65536 elements truncated to 0 and silently produced an EMPTY array
+        // (array_length -> NULL) instead of the right answer.
+        nelems: u32,
         frame: u32,
         elmtype: Oid,
         elmlen: i16,
@@ -157,7 +161,10 @@ pub enum Step {
     // anonymous-RECORD tupdesc, arena-lived for the plan.
     RowExprStep {
         elems: NonNull<NullableDatum>,
-        nelems: u16,
+        // C rowexpr.nelems is `int` too (see ArrayExprStep above). In practice
+        // the blessed tupdesc bounds this well below 65535, but the field must
+        // not be the thing that silently wraps.
+        nelems: u32,
         frame: u32,
         desc: NonNull<::types_tuple::TupleDescData<'static>>,
         out: OutRef,
@@ -262,7 +269,7 @@ pub enum Step {
     HstoreSbsrefAssign { state: NonNull<crate::hstoresubs::HstoreSbsState>, out: OutRef },
     // slots: nelems compile-allocated NullableDatum arg targets (C's
     // d.minmax.values/nulls); call is the type's btree cmp proc.
-    MinMax { call: FuncCall, slots: NonNull<NullableDatum>, nelems: u16, least: bool, out: OutRef },
+    MinMax { call: FuncCall, slots: NonNull<NullableDatum>, nelems: u32, least: bool, out: OutRef },
     NextValueExpr { seqid: Oid, seqtypid: Oid, out: OutRef },
     // C EEOP_JSON_CONSTRUCTOR: arg subexprs write jcstate's slots; constant
     // metadata + split scratch behind one plan-mcx pointer.
